+++ warthog/__init__.py
@@ -14,4 +14,4 @@ warthog
 Simple client for A10 load balancers.
 """
 
-__version__ = '0.9.0'
+__version__ = '0.9.1'